refactor(tests): split task tests into per-task files
Splits the monolithic task_execute_test.go and integration_test.go
into per-task files mirroring the existing <task>_task.go convention.
Per-task parsing tests move out of main_test.go into the same
per-task files. Shared integration helpers and TestMain move to
integration_helpers_test.go. Generic GetTasks parser tests and
registry/meta tests stay in main_test.go; cross-task workflow
integration tests stay in integration_test.go. No production code
changes; baseline of 165 Test functions, 127 unit pass, and 31
integration tests is preserved.
