+++ tasks/checks_toggle_task_test.go
@@ -1,0 +1,13 @@
+package tasks
+
+import (
+	"testing"
+)
+
+func TestChecksToggleTaskInvalidState(t *testing.T) {
+	task := ChecksToggleTask{App: "test-app", State: "invalid"}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("Execute with invalid state should return an error")
+	}
+}
