+++ tasks/nginx_property_task_test.go
@@ -1,0 +1,71 @@
+package tasks
+
+import (
+	"strings"
+	"testing"
+)
+
+func TestNginxPropertyTaskInvalidState(t *testing.T) {
+	task := NginxPropertyTask{App: "test-app", Property: "proxy-read-timeout", State: "invalid"}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("Execute with invalid state should return an error")
+	}
+}
+
+func TestNginxPropertyTaskMissingApp(t *testing.T) {
+	task := NginxPropertyTask{Property: "proxy-read-timeout", Value: "120s", State: StatePresent}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("Execute without app and global=false should return an error")
+	}
+}
+
+func TestNginxPropertyTaskGlobalWithAppSet(t *testing.T) {
+	task := NginxPropertyTask{
+		App:      "test-app",
+		Global:   true,
+		Property: "proxy-read-timeout",
+		Value:    "120s",
+		State:    StatePresent,
+	}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("expected error when both global and app are set")
+	}
+	if !strings.Contains(result.Error.Error(), "must not be set when 'global' is set to true") {
+		t.Errorf("unexpected error: %v", result.Error)
+	}
+}
+
+func TestNginxPropertyTaskPresentWithoutValue(t *testing.T) {
+	task := NginxPropertyTask{
+		App:      "test-app",
+		Property: "proxy-read-timeout",
+		Value:    "",
+		State:    StatePresent,
+	}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("expected error when present state has no value")
+	}
+	if !strings.Contains(result.Error.Error(), "invalid without a value") {
+		t.Errorf("unexpected error: %v", result.Error)
+	}
+}
+
+func TestNginxPropertyTaskAbsentWithValue(t *testing.T) {
+	task := NginxPropertyTask{
+		App:      "test-app",
+		Property: "proxy-read-timeout",
+		Value:    "120s",
+		State:    StateAbsent,
+	}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("expected error when absent state has a value")
+	}
+	if !strings.Contains(result.Error.Error(), "invalid with a value") {
+		t.Errorf("unexpected error: %v", result.Error)
+	}
+}
