+++ tasks/storage_ensure_task_test.go
@@ -1,0 +1,50 @@
+package tasks
+
+import (
+	"testing"
+)
+
+func TestStorageEnsureTaskInvalidState(t *testing.T) {
+	task := StorageEnsureTask{App: "test-app", Chown: "heroku", State: "invalid"}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("Execute with invalid state should return an error")
+	}
+}
+
+func TestStorageEnsureValidChownValues(t *testing.T) {
+	validValues := []string{"heroku", "herokuish", "paketo", "root", "false"}
+	for _, chown := range validValues {
+		task := StorageEnsureTask{App: "test-app", Chown: chown, State: StatePresent}
+		result := task.Execute()
+		// These will fail because dokku isn't running, but should NOT fail
+		// due to invalid chown value
+		if result.Error != nil && result.Error.Error() == "invalid chown value specified" {
+			t.Errorf("chown value %q should be valid but was rejected", chown)
+		}
+	}
+}
+
+func TestStorageEnsureInvalidChownValue(t *testing.T) {
+	task := StorageEnsureTask{App: "test-app", Chown: "packeto", State: StatePresent}
+	result := task.Execute()
+	if result.Error == nil || result.Error.Error() != "invalid chown value specified" {
+		t.Errorf("chown value 'packeto' (misspelled) should be rejected as invalid")
+	}
+}
+
+func TestStorageEnsureAbsentStateReturnsError(t *testing.T) {
+	task := StorageEnsureTask{App: "test-app", Chown: "heroku", State: StateAbsent}
+	result := task.Execute()
+	if result.Error == nil {
+		t.Fatal("Execute with absent state should return an error for storage ensure")
+	}
+}
+
+func TestStorageEnsureEmptyChown(t *testing.T) {
+	task := StorageEnsureTask{App: "test-app", Chown: "", State: StatePresent}
+	result := task.Execute()
+	if result.Error == nil || result.Error.Error() != "invalid chown value specified" {
+		t.Errorf("expected 'invalid chown value specified' error, got: %v", result.Error)
+	}
+}
